# Conflicts:
#	src/main/java/org/sjk/controller/LoginPageController.java
#	src/main/webapp/WEB-INF/jsp/ip_error.jsp
#	src/main/webapp/WEB-INF/jsp/user_login.jsp

+++ src/main/java/org/sjk/dao/ActionDao.java
@@ -8,7 +8,9 @@ import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import javax.sql.DataSource;
+import java.sql.Time;
 import java.sql.Timestamp;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -28,39 +30,34 @@ public class ActionDao {
                 "a_u_id bigint not null, " +
                 "a_i_id bigint not null, " +
                 "a_czas_akcji timestamp not null);";
-        System.out.println("\n"+initTableScript.toUpperCase()+"\n");
         jdbcTemplate.execute(initTableScript);
     }
 
     public void updateRelations() {
         String updateActionTableScript="alter table Akcje add foreign key (a_u_id) references Uzytkownicy(u_id);\n" +
                 "alter table Akcje add foreign key (a_i_id) references IP(i_id);";
-        System.out.println("\n"+updateActionTableScript.toUpperCase()+"\n");
         jdbcTemplate.execute(updateActionTableScript);
     }
 
     public List<Action> findAllActions(long userId){
-        String findActionsForUserScript="select * from Akcje where a_u_id=?";
-        System.out.println("\n"+findActionsForUserScript.toUpperCase()+"\n");
-        List<Action> userActions=jdbcTemplate.query(findActionsForUserScript,new BeanPropertyRowMapper<Action>(Action.class),userId);
+        String findActionsForUserScript="select * from Akcje where a_u_id="+userId;
+        List<Action> userActions=jdbcTemplate.query(findActionsForUserScript,new BeanPropertyRowMapper<Action>(Action.class));
         return userActions;
     }
 
     public void removeActionForUser(String actionName, Timestamp actionTime, long userId){
-        String removeUserActionScript="delete from Akcje where a_nazwa=? and a_czas_akcji=? and a_u_id=?";
-        System.out.println("\n"+removeUserActionScript.toUpperCase()+"\n");
-        jdbcTemplate.update(removeUserActionScript,new Object[]{actionName,actionTime,userId});
+        String removeUserActionScript="delete from Akcje where a_nazwa=\'"+actionName+"\' and a_czas_akcji=\'"+actionTime+
+                "\' and a_u_id="+userId;
+        jdbcTemplate.update(removeUserActionScript);
     }
 
-    public void removeActionsForUser(long userId){
-        String removeUserActionsScript="delete from Akcje where a_u_id=?";
-        System.out.println("\n"+removeUserActionsScript.toUpperCase()+"\n");
-        jdbcTemplate.update(removeUserActionsScript,new Object[]{userId});
+    public void removeActionsForUser(long userId) {
+        String removeUserActionsScript = "delete from Akcje where a_u_id=" + userId;
+        jdbcTemplate.update(removeUserActionsScript);
     }
-
     public void insertAction(String actionName, Timestamp actionTime, long userId, long ipId){
-        String insertActionScript="insert into Akcje(a_nazwa,a_czas_akcji,a_u_id,a_i_id) values(?,?,?,?);";
-        System.out.println("\n"+insertActionScript.toUpperCase()+"\n");
-        jdbcTemplate.update(insertActionScript,new Object[]{actionName,actionTime,userId,ipId});
+        String insertActionScript="insert into Akcje(a_nazwa,a_czas_akcji,a_u_id,a_i_id) values(\'"+actionName+
+                "\', \'"+actionTime+"\',"+userId+","+ipId+")";
+        jdbcTemplate.update(insertActionScript);
     }
 }
